requirements.txt
anyio==4.2.0
boltons @ file:///work/ci_py311/boltons_1677685195580/work
brotlipy==0.7.0
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi @ file:///work/ci_py311/cffi_1676822533496/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
conda @ file:///croot/conda_1696257509808/work
conda-content-trust @ file:///croot/conda-content-trust_1693490622020/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1696257536937/work/src
conda-package-handling @ file:///croot/conda-package-handling_1690999929514/work
conda_package_streaming @ file:///croot/conda-package-streaming_1690987966409/work
cryptography @ file:///croot/cryptography_1694444244250/work
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna @ file:///work/ci_py311/idna_1676822698822/work
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
libmambapy @ file:///croot/mamba-split_1694187754698/work/libmambapy
packaging @ file:///croot/packaging_1693575174725/work
pluggy @ file:///work/ci_py311/pluggy_1676822818071/work
pycosat @ file:///croot/pycosat_1696536503704/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-telegram-bot==20.7
requests @ file:///croot/requests_1690400202158/work
ruamel.yaml @ file:///work/ci_py311/ruamel.yaml_1676838772170/work
sniffio==1.3.0
tqdm @ file:///croot/tqdm_1679561862951/work
truststore @ file:///croot/truststore_1695244293384/work
urllib3 @ file:///croot/urllib3_1686163155763/work
zstandard @ file:///work/ci_py311_2/zstandard_1679339489613/work
